+++ CHANGELOG.md
@@ -1,8 +1,16 @@
-v1.1 -- July 21, 2012
+## v2.0 - December 28, 2014
+
+- Refactor system structure
+- Use Symfony's console component.
+- Add unit tests
+- Use a shell wrapper for mocking
+- Remove Linux support for now
+
+## v1.1 - July 21, 2012
 
 - Created a Benchmarking class
 
-v1.0 -- July 20, 2012
+## v1.0 - July 20, 2012
 
 - Show PID on script start-up
 - Tested script performance
@@ -10,11 +18,11 @@ v1.0 -- July 20, 2012
 - Fixed Linux header issues
 - Fixed Linux notification errors
 
-v0.2.0 -- July 19, 2012
+## v0.2.0 - July 19, 2012
 
 - Added the notification system for Linux and OS X Systems
 
-v0.1.0 -- July 19, 2012
+## v0.1.0 - July 19, 2012
 
 - Trace user changes [ Connection and Disconnection ]
 - Detect all the headers from the `w` command